Use GSRoot Overridden Allocators Consistently
Many API headers already include GSRoot.hpp (and thus GSNew and
GSMalloc), but the headers have to be included everywhere to ensure
that the Add-On consistently uses new/delete and malloc/free.
This is especially important for the malloc/free override macros, which
need to be present in all translation units.