Embed version metadata in addons on all platforms
With this change, the add-on tools will rely more on information
provided by the add-on template's config.json file for improved UX.

Add-ons on both platforms will get their version information from the
config.json file's "version" field, which must be a 1, 2 or 3 component
version number in the form on "123", "1.23" or "1.2.3". Components that
aren't provided will default to 0, i.e. "123.0.0" with a one component
version number.

The "copyright" field of config.json is an object with fields "name" and
"year", which are used to fill in company and copyright metadata where
applicable. The "year" field is a format string that will be used as an
input to the string(TIMESTAMP) CMake command. For details on what format
options are available, see the documentation:

https://cmake.org/cmake/help/latest/command/string.html#timestamp

These new "version" and "copyright" fields are required.

Some platform specific information:

For Windows: There was previously no VERSIONINFO resource generated for
addons, so this is completely new.

For MacOS: Some property list items were removed, replaced and moved
around to better reflect the documentation from Apple and to get rid of
obsolete and unused items.