Embed AC version(s) in root product version
Since the root versions require their components to be in the 16-bit
unsigned integer range (0-65535), I have decided to move the AC version
and the GS build number into the root product version. The root product
version is not displayed in the Detail tab in Properties if you rename
the add-on to .dll, because the "StringFileInfo.ProductVersion" value
takes precedence.

Making the AC version the 2nd and the GS build number the 4th component
will make it very easy to read those versions from Archicad later. See
the "dwProductVersionMS" and "dwProductVersionLS" members:

https://learn.microsoft.com/en-us/windows/win32/api/verrsrc/ns-verrsrc-vs_fixedfileinfo